+++ scratch.js
@@ -2,18 +2,46 @@
 import nlp from './src/three.js'
 import plg from './plugins/dates/src/plugin.js'
 nlp.plugin(plg)
-
-nlp.verbose('tagger')
+// nlp.verbose('tagger')
 
 let txt = ''
 let doc
-let m
+// let m
 
-// doc = nlp('Those are Great Danes')
-// doc.nouns(0).toSingular()
-// console.log(doc.text())
 
-doc = nlp('30/01/194').debug()
+// bug 1
+// doc = nlp('we swim')
+// console.log(doc.verbs().conjugate())
 
-// console.log(doc.world)
-// console.log(nlp.world())
+// bug 2
+// doc = nlp('blew').debug()
+// console.log(doc.verbs().conjugate())
+
+nlp.plugin({
+  irregulars: {
+    get: {
+      pastTense: 'gotten',
+      presentTense: 'getts',
+      gerund: 'gettin'
+    },
+    sly: {
+      comparative: 'slyer',
+      superlative: 'sliiest',
+    }
+  }
+})
+doc = nlp('gettin').tag('Gerund').debug()
+console.log(doc.verbs().conjugate())
+// console.log(doc.verbs().toGerund().text())
+
+
+// date issues:
+// 'the month before christmas' vs 'a month before christmas'
+// middle september
+// end of september
+// first half of march
+// week of june 3rd
+// fridays in june
+// every saturday
+// now
+// until christmas