+++ src/2-two/preTagger/model/clues/index.js
@@ -9,6 +9,7 @@ import personNoun from './person-noun.js'
 import personVerb from './person-verb.js'
 import personPlace from './person-place.js'
 import unitNoun from './unit-noun.js'
+import honorificNoun from './honorific-noun.js'
 
 const clues = {
   'Adj|Gerund': adjGerund,
@@ -22,6 +23,7 @@ const clues = {
   'Person|Verb': personVerb,
   'Person|Place': personPlace,
   'Unit|Noun': unitNoun,
+  'Honorific|Noun': honorificNoun
 }
 
 const copy = (obj, more) => {
@@ -34,9 +36,7 @@ const copy = (obj, more) => {
 
 // make a copy of this one
 clues['Plural|Verb'] = {
-  beforeWords: copy(clues['Noun|Verb'].beforeWords, {
-
-  }),
+  beforeWords: copy(clues['Noun|Verb'].beforeWords, {}),
   afterWords: copy(clues['Noun|Verb'].afterWords, {
     his: 'PresentTense', her: 'PresentTense', its: 'PresentTense',
     in: null, to: null,