+++ src/3-three/misc/adjectives/index.js
@@ -39,7 +39,19 @@ const api = function (View) {
     adverbs() {
       return this.before('#Adverb+$').concat(this.after('^#Adverb+'))
     }
-
+    conjugate(n) {
+      const { adjToComparative, adjToSuperlative, adjToNoun, adjToAdverb } = this.methods.two.transform
+      return getNth(this, n).map(adj => {
+        let root = toRoot(adj)
+        return {
+          Adjective: root,
+          Comparative: adjToComparative(root, this.model),
+          Superlative: adjToSuperlative(root, this.model),
+          Noun: adjToNoun(root, this.model),
+          Adverb: adjToAdverb(root, this.model),
+        }
+      }, [])
+    }
     toComparative(n) {
       const { adjToComparative } = this.methods.two.transform
       return getNth(this, n).map(adj => {