+++ src/API/extend.js
@@ -31,6 +31,54 @@ function mergeQuick(model, plugin) {
   return model
 }
 
+const addIrregulars = function (model, conj) {
+  let m = model.two.models || {}
+  Object.keys(conj).forEach(k => {
+    // verb forms
+    if (conj[k].pastTense) {
+      if (m.toPast) {
+        m.toPast.exceptions[k] = conj[k].pastTense
+      }
+      if (m.fromPast) {
+        m.fromPast.exceptions[conj[k].pastTense] = k
+      }
+    }
+    if (conj[k].presentTense) {
+      if (m.toPresent) {
+        m.toPresent.exceptions[k] = conj[k].presentTense
+      }
+      if (m.fromPresent) {
+        m.fromPresent.exceptions[conj[k].presentTense] = k
+      }
+    }
+    if (conj[k].gerund) {
+      if (m.toGerund) {
+        m.toGerund.exceptions[k] = conj[k].gerund
+      }
+      if (m.fromGerund) {
+        m.fromGerund.exceptions[conj[k].gerund] = k
+      }
+    }
+    // adjective forms
+    if (conj[k].comparative) {
+      if (m.toComparative) {
+        m.toComparative.exceptions[k] = conj[k].comparative
+      }
+      if (m.fromComparative) {
+        m.fromComparative.exceptions[conj[k].comparative] = k
+      }
+    }
+    if (conj[k].superlative) {
+      if (m.toSuperlative) {
+        m.toSuperlative.exceptions[k] = conj[k].superlative
+      }
+      if (m.fromSuperlative) {
+        m.fromSuperlative.exceptions[conj[k].superlative] = k
+      }
+    }
+  })
+}
+
 const extend = function (plugin, world, View, nlp) {
   const { methods, model, compute, hooks } = world
   if (plugin.methods) {
@@ -38,6 +86,9 @@ const extend = function (plugin, world, View, nlp) {
   }
   if (plugin.model) {
     mergeDeep(model, plugin.model)
+  }
+  if (plugin.irregulars) {
+    addIrregulars(model, plugin.irregulars)
   }
   // shallow-merge compute
   if (plugin.compute) {