+++ changelog.md
@@ -8,6 +8,14 @@ While all _Major_ releases should be reviewed, our only _large_ releases are **v
 
 <!-- #### 14.5.0 [Unreleased]
 -->
+
+#### 14.5.0 [August 2022]
+- **[fix]** - possible runtime error in setTag method
+- **[change]** - make #Honorific always a #Person #951
+- **[new]** - manually change conjugations/inflections from plugin #949
+- **[new]** - `.adjectives().conjugate()` method
+- **[update]** - dependencies
+
 #### 14.4.5 [August 2022]
 - **[fix]** - fix logic for greedy-negative matches - #936 
 - **[fix]** - fix tagging for 3-digit year iso dates - #868